backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
httpx==0.28.1
tzdata==2025.2
pillow==11.2.1
websockets==13.1
paho-mqtt==2.1.0
python-multipart==0.0.20
pymysql==1.1.1
